IRIS_LARP/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
sqlalchemy>=2.0.25
jinja2>=3.1.3
openai>=1.10.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.26.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7
python-jose[cryptography]>=3.3.0
pytest>=7.4.3
